amdxdna: unify dev_info in drv.h for PCI and AUX builds
Move struct amdxdna_dev_info and struct amdxdna_fw_feature_tbl from
amdxdna_pci_drv.h into amdxdna_drv.h and drop the duplicate definition
from amdxdna_aux_drv.h.
Shared sources (e.g. ve2_aux.c via gem/ctx headers) include the PCI
header while AUX also pulls aux_drv.h, which caused a redefinition error
when XDNA_BUS_TYPE=aux. One layout keeps AUX partial initializers valid
(zero-filled unused PCI fields).
