+++ drivers/accel/amdxdna/amdxdna_aux_drv.h
@@ -9,17 +9,4 @@
 #include "amdxdna_drv.h"
 #include "drm/amdxdna_accel.h"
 
-/*
- * struct amdxdna_dev_info - Device hardware information
- * Record device static information, like reg, mbox, PSP, SMU bar index
- */
-struct amdxdna_dev_info {
-	int				device_type;
-	const struct amdxdna_dev_priv	*dev_priv;
-	const struct amdxdna_dev_ops	*ops;
-};
-
-/* Add device info below */
-extern const struct amdxdna_dev_info dev_ve2_info;
-
 #endif /* _AMDXDNA_AUX_DRV_H_ */
