+++ drivers/accel/amdxdna/amdxdna_pci_drv.h
@@ -8,36 +8,6 @@
 
 #include "amdxdna_drv.h"
 
-struct amdxdna_fw_feature_tbl {
-	u64 features;
-	u32 major;
-	u32 max_minor;
-	u32 min_minor;
-};
-
-/*
- * struct amdxdna_dev_info - Device hardware information
- * Record device static information, like reg, mbox, PSP, SMU bar index
- */
-struct amdxdna_dev_info {
-	int				reg_bar;
-	int				mbox_bar;
-	int				sram_bar;
-	int				psp_bar;
-	int				smu_bar;
-	int				device_type;
-	int				first_col;
-	u32				dev_mem_buf_shift;
-	u64				dev_mem_base;
-	size_t				dev_mem_size;
-	const char			*default_vbnv;
-	const struct amdxdna_rev_vbnv	*rev_vbnv_tbl;
-	size_t				dev_heap_max_size;
-	const struct amdxdna_dev_priv	*dev_priv;
-	const struct amdxdna_fw_feature_tbl *fw_feature_tbl;
-	const struct amdxdna_dev_ops	*ops;
-};
-
 /* Add device info below */
 extern const struct amdxdna_dev_info dev_npu1_info;
 extern const struct amdxdna_dev_info dev_npu3_pf_info;
